feat: Add version flag and build configuration
This commit introduces a version flag to the application, allowing users
to easily check the current version.

It also adds a build configuration to set the version during the build
process using `-ldflags`.

This change ensures the application version is displayed when using the
version flag and ensures error returns in some scenarios.

Additionally, start.sh was added to the gitignore file.